Move MathError definition
